Small updates
Adds a compile-time define `COMPACT_UPDATE` which if set uses the
original 1-line update statement that uses array syntax. Otherwise
new code that uses an explicit do loop for the update is used. NAG
performed very poorly with the original code, and Intel and gfortran
perform equally well with either.